% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in var.R
\name{bc_variance}
\alias{bc_variance}
\title{Business Cycle Frequency Variance}
\usage{
bc_variance(irf_array)
}

\arguments{
\item{irf_array}{3D array of IRFs (n_hor x n_var x n_var)}
}

\value{
Covariance matrix
}

\description{
Computes variance at business cycle frequencies (6-32 quarters).
}

