% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmw26-package.R
\docType{package}
\name{cmw26-package}
\alias{cmw26}
\alias{cmw26-package}
\title{cmw26: Evaluating Monetary Policy Counterfactuals}
\description{
R port of the replication files for "Evaluating Monetary Policy Counterfactuals: (When) Do We Need Structural Models?" by Caravello, McKay and Wolf. Implements Bayesian VAR estimation, impulse response analysis, and frequency domain variance decomposition using Armadillo for efficient matrix computations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cmw26}
  \item Report bugs at \url{https://github.com/pachadotdev/cmw26/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Tomas Caravello
  \item Alisdair McKay
  \item Christian Wolf
}

}
\keyword{internal}
