% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cnfctl.R
\name{compute_cnfctl_wold}
\alias{compute_cnfctl_wold}
\title{Counterfactual Wold IRFs for all shocks}
\usage{
compute_cnfctl_wold(
  wold_base,
  Pi_m_draws,
  Y_m_draws,
  I_m_draws,
  rule = NULL,
  use_optpol = FALSE,
  weights = NULL
)
}

\arguments{
\item{wold_base}{3D array (n_var x n_shocks x T) of baseline Wold IRFs}

\item{Pi_m_draws}{List of inflation policy IRFs, one per draw}

\item{Y_m_draws}{List of output policy IRFs}

\item{I_m_draws}{List of interest rate policy IRFs}

\item{rule}{Policy rule specification}

\item{use_optpol}{If TRUE, use optimal policy; otherwise use simple rule}

\item{weights}{Optimal policy weights (required if use_optpol = TRUE)}
}

\value{
4D array (n_var x n_shocks x T x n_draws) of counterfactual Wold IRFs
}

\description{
Given baseline Wold IRFs and policy shock IRFs, compute counterfactual
 Wold IRFs under a policy rule.
}

