% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in var.R
\name{compute_irf}
\alias{compute_irf}
\title{IRF from State-Space Model}
\usage{
compute_irf(ME, MX, MY, n_periods = 20, shock_idx = 1)
}

\arguments{
\item{ME}{Impact matrix (nx x ne)}

\item{MX}{State transition matrix (nx x nx)}

\item{MY}{Observation matrix (ny x nx)}

\item{n_periods}{Number of periods}

\item{shock_idx}{Which shock to compute IRF for (1-indexed)}
}

\value{
A list with YY (observations) and XX (states) IRFs
}

\description{
Compute IRFs from state-space representation of a model.
}

