% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{diff_vec}
\alias{diff_vec}
\title{Difference a vector}
\usage{
diff_vec(x, k = 1)
}

\arguments{
\item{x}{Numeric vector}

\item{k}{Difference order}
}

\value{
Differenced vector (length n - k)
}

\description{
Compute first difference.
}

