% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in var.R
\name{freq_variance}
\alias{freq_variance}
\title{Frequency Domain Variance}
\usage{
freq_variance(irf_array, omega_low, omega_high, n_grid = 100)
}

\arguments{
\item{irf_array}{3D array of IRFs (n_hor x n_var x n_var)}

\item{omega_low}{Lower frequency bound}

\item{omega_high}{Upper frequency bound}

\item{n_grid}{Number of grid points for integration}
}

\value{
Covariance matrix
}

\description{
Integrates the spectral density of a VAR over a frequency band.
}

