% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{hamilton}
\alias{hamilton}
\title{Hamilton filter}
\usage{
hamilton(y, h = 8, p = 4)
}

\arguments{
\item{y}{Numeric vector}

\item{h}{Forecast horizon (default 8 for quarterly data)}

\item{p}{Number of lags (default 4)}
}

\value{
Detrended series with leading NAs
}

\description{
Remove business cycle component using Hamilton (2018) filter.
}

