% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{hp_filter}
\alias{hp_filter}
\title{HP filter}
\usage{
hp_filter(x, lambda = 1600)
}

\arguments{
\item{x}{Numeric vector}

\item{lambda}{Smoothing parameter (1600 for quarterly, 100 for annual)}
}

\value{
A list with trend and cycle components
}

\description{
Hodrick-Prescott filter for trend extraction.
}

