% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{kde}
\alias{kde}
\title{Kernel density estimation}
\usage{
kde(x, xi, h = 0, kernel = "gaussian")
}

\arguments{
\item{x}{Evaluation points}

\item{xi}{Data points}

\item{h}{Bandwidth (0 for automatic selection)}

\item{kernel}{Type: "gaussian" or "epanechnikov"}
}

\value{
Density estimates at x
}

\description{
Estimate probability density function using kernel methods.
}

