% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{lag_vec}
\alias{lag_vec}
\title{Lag a vector}
\usage{
lag_vec(x, k = 1)
}

\arguments{
\item{x}{Numeric vector}

\item{k}{Number of lags (positive) or leads (negative)}
}

\value{
Lagged vector
}

\description{
Create lagged values, filling with NA.
}

