% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{sample_from_models}
\alias{sample_from_models}
\title{Extract draws from model collection}
\usage{
sample_from_models(n_draws, model_probs, Pi_m_list, Y_m_list, I_m_list)
}

\arguments{
\item{n_draws}{Number of draws}

\item{model_probs}{Model probability weights}

\item{Pi_m_list}{List of lists: Pi_m_list[[model]][[draw]]}

\item{Y_m_list}{List of lists for output IRFs}

\item{I_m_list}{List of lists for interest rate IRFs}
}

\value{
A list with sampled Pi_m, Y_m, I_m draws
}

\description{
Sample policy IRF draws from a collection of models according to posterior
 probabilities.
}

