% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{sample_models_r}
\alias{sample_models_r}
\title{Sample from models with probability weights}
\usage{
sample_models_r(n, probs)
}

\arguments{
\item{n}{Number of samples}

\item{probs}{Probability weights (will be normalized)}
}

\value{
Vector of model indices (1-indexed)
}

\description{
Draw model indices according to posterior probabilities.
}

