% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cnfctl.R
\name{vma_cov}
\alias{vma_cov}
\title{VMA-implied covariance matrix from Wold IRFs}
\usage{
vma_cov(wold)
}

\arguments{
\item{wold}{3D array (n_var x n_shocks x T) of Wold IRFs}
}

\value{
n_var x n_var covariance matrix
}

\description{
Computes the covariance matrix implied by the VMA representation of Wold IRFs.
}

