% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annual_avg}
\alias{annual_avg}
\title{Annual average of quarterly data}
\usage{
annual_avg(x, start_quarter = 1)
}
\arguments{
\item{x}{Numeric vector of quarterly data}

\item{start_quarter}{Starting quarter (1-4)}
}
\value{
Vector of annual averages
}
\description{
Compute annual averages from quarterly observations.
}
