% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{compute_var_irf}
\alias{compute_var_irf}
\title{Impulse Response Functions from VAR}
\usage{
compute_var_irf(bvar_result, n_var, n_lags, n_hor = 20, use_draws = FALSE)
}
\arguments{
\item{bvar_result}{Result from estimate_bvar()}

\item{n_var}{Number of variables}

\item{n_lags}{Number of lags}

\item{n_hor}{Horizon for IRFs}

\item{use_draws}{If TRUE, compute IRFs for all posterior draws}
}
\value{
A list of IRF matrices or (if use_draws) a 4D array
}
\description{
Compute IRFs from VAR coefficient matrices and covariance.
}
