% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnfctl.R
\name{cov_to_corr_r}
\alias{cov_to_corr_r}
\title{Correlation matrix from covariance}
\usage{
cov_to_corr_r(cov)
}
\arguments{
\item{cov}{Covariance matrix}
}
\value{
Correlation matrix
}
\description{
Converts a covariance matrix to a correlation matrix.
}
