% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detrend}
\alias{detrend}
\title{Detrend a time series}
\usage{
detrend(y, const_type = 1)
}
\arguments{
\item{y}{Numeric vector or matrix}

\item{const_type}{1 = constant only, 2 = constant + linear trend}
}
\value{
A list with Beta (trend coefficients) and Res (detrended series)
}
\description{
Remove deterministic trend using OLS.
}
