% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{growth_rate}
\alias{growth_rate}
\title{Growth rate}
\usage{
growth_rate(x, k = 1, annualize = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{k}{Period for growth rate}

\item{annualize}{If TRUE, annualize the growth rate}
}
\value{
Growth rate vector
}
\description{
Compute percentage growth rate.
}
