% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_along}
\alias{quantile_along}
\title{Quantile summary for arrays}
\usage{
quantile_along(x, probs = c(0.16, 0.5, 0.84), along = length(dim(x)))
}
\arguments{
\item{x}{Array}

\item{probs}{Quantile probabilities}

\item{along}{Which dimension to summarize over}
}
\value{
Array with quantiles
}
\description{
Compute quantiles along a specified dimension.
}
