% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quarterly_dates}
\alias{quarterly_dates}
\title{Create quarterly date sequence}
\usage{
quarterly_dates(start_year, start_quarter, n)
}
\arguments{
\item{start_year}{Starting year}

\item{start_quarter}{Starting quarter}

\item{n}{Number of quarters}
}
\value{
Vector of Date objects (first day of each quarter)
}
\description{
Generate a sequence of dates for quarterly data.
}
