% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnfctl.R
\name{set_optpol_weights}
\alias{set_optpol_weights}
\title{Set up optimal policy loss function weights}
\usage{
set_optpol_weights(
  T,
  lambda_pi = 1,
  lambda_y = 1,
  lambda_i = 0,
  lambda_di = 1,
  beta = 1
)
}
\arguments{
\item{T}{Number of periods}

\item{lambda_pi}{Weight on inflation squared}

\item{lambda_y}{Weight on output gap squared}

\item{lambda_i}{Weight on interest rate squared}

\item{lambda_di}{Weight on interest rate change squared}

\item{beta}{Discount factor}
}
\value{
A list with W_pi, W_y, W_i (T x T matrices)
}
\description{
Creates W_pi, W_y, W_i weight matrices for optimal policy computation.
}
