% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var_fcst}
\alias{var_fcst}
\title{VAR Forecasts}
\usage{
var_fcst(
  data,
  n_lags,
  constant,
  B,
  fcst_date,
  fcst_length,
  include_current = FALSE
)
}
\arguments{
\item{data}{A T x n_var matrix of data}

\item{n_lags}{Number of lags}

\item{constant}{Deterministic terms (0, 1, 2, or 3)}

\item{B}{Coefficient matrix from VAR estimation}

\item{fcst_date}{Forecast origin (row index in original data)}

\item{fcst_length}{Number of periods to forecast}

\item{include_current}{If TRUE, include current period in forecast}
}
\value{
A matrix of forecasts (fcst_length x n_var)
}
\description{
Generate forecasts from a VAR model.
}
