% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize a vector}
\usage{
winsorize(x, w = 90)
}
\arguments{
\item{x}{Numeric vector}

\item{w}{Winsorization percentage (default 90, meaning trim 5 percent from each tail)}
}
\value{
Winsorized vector
}
\description{
Replace extreme values with specified percentiles.
}
\examples{
x <- c(1, 2, 3, 100, 4, 5)
winsorize(x, 80)
}
